Refactoring: extract Registry
